metadata: mp3: support for files longer than 2Gb
This still needs support for 64bit seek:
If you are compiling for a 32-bit embedded system, standard `lseek` uses a 32-bit signed `off_t`. The check you provided will work fine for seeking to `0`, but if you later try to seek to an offset > 2.14 GB, standard `lseek` will fail, return `-1`, and set errno to `EOVERFLOW`. To seek past 2 GB, you must compile with Large File Support enabled (e.g., passing `-D_FILE_OFFSET_BITS=64` to gcc).

+++ apps/plugin.h
@@ -783,7 +783,7 @@ struct plugin_api {
     /* metadata */
     bool (*get_metadata)(struct mp3entry* id3, int fd, const char* trackname);
     const char* (*get_codec_string)(int codectype);
-    int (*count_mp3_frames)(int fd,  int startpos,  int filesize,
+    int (*count_mp3_frames)(int fd,  int64_t startpos,  int64_t filesize,
                      void (*progressfunc)(int),
                      unsigned char* buf, size_t buflen);
     int (*create_xing_header)(int fd, long startpos, long filesize,
